fix(brett): extract undo-redo UI + startIdleSweep to resolve S1 gate [T000470]
- board-boot.ts: extract initUndoRedo() to ui/undo-redo-ui.ts (622→446 lines)
- ws-handler.ts: move startIdleSweep to ws-admin-commands.ts (602→578 lines)
- Both files now under the 600-line S1 gate
